Changed orbit.py to pygame_orbit.py, and added file header description
