+++ README.md
@@ -1,9 +1,9 @@
 # Pygame-Orbit
 
 #### Description:
-Pygame-Orbit is a pygame demo, that displays a 2d solar system.  The solar system contains a star, a planet orbiting the star, and several moons orbiting the planet or other moons.
+pygame_orbit is a pygame demo, that displays a 2d solar system.  The solar system contains a star, a planet orbiting the star, and several moons orbiting the planet or other moons.
 
-The planets and moons can be selected which will highlight them in blue and also draw a line between them and the object they are orbiting.
+The planets and moons can be selected, which will highlight them in blue and also draw a line between them and the object they are orbiting.
 
 #### Requirements:
 - [Python](https://www.python.org/) 2 or 3
